refactor: rename client.audio.speech to client.audio.speech.create
