+++ examples/coze-js-node/src/voice.mjs
@@ -25,7 +25,7 @@ try {
 }
 
 try {
-  const speechBuffer = await client.audio.speech({
+  const speechBuffer = await client.audio.speech.create({
     input:
       '今天天气真是太好了，阳光灿烂，心情超级棒，但是朋友最近的感情问题也让我心痛不已，好像世界末日一样，真的好为他难过。',
     voice_id: '742894224871836***',
@@ -50,13 +50,11 @@ async function listAllVoices() {
     let allVoices = [];
 
     while (true) {
-      console.log('page', page);
       const response = await client.audio.voices.list({
         page_size: PAGE_SIZE,
         page_num: page,
       });
       allVoices = allVoices.concat(response.voice_list);
-      console.log('response', response);
       if (response.has_more === false) {
         break;
       }
