+++ packages/coze-js/src/resources/audio/speech/index.ts
@@ -1,0 +1,1 @@
+export * from './speech.js';
